move single module
